second commit

+++ src/controllers/user.py
@@ -1,54 +1,71 @@
 from flask import request, jsonify, Response
 from bson import json_util, ObjectId
 
+from flask_jwt_extended import jwt_required, get_jwt_identity
+
 from config.database import mongo
 
 def createUser(): #Add user
     try:
-
         fullname = request.json['fullname']
+        username = request.json['username']
+        password = request.json['password']
 
-        existe = mongo.db.user.find_one({'fullname': fullname})
+        isRegister = mongo.db.user.find_one({'username': username})
 
-        if existe:
-            return jsonify({'msg': f'User {existe["fullname"]} added previously'}), 400
+        if isRegister:
+            return jsonify({'msg': f'User {isRegister["username"]} added previously'}), 400
 
-        if fullname:
-            user = mongo.db.user.insert_one({ 'fullname': fullname })
+        user = mongo.db.user.insert_one({
+            'fullname': fullname, 'username': username, "password": password
+        })
 
-            return jsonify({'msg': f'User {str(user.inserted_id)} added successfully'}), 200
+        return jsonify({'msg': f'User {str(user.inserted_id)} added successfully'}), 200
     except Exception as e:
-        return jsonify(({ 'msg': f'Error {str(e)}' })), 500
+        return jsonify({ 'msg': f'Error {str(e)}' }), 500
 
 
-def readUser(): #Get users
+def findAllUser(): #Get users
     try:
         data = mongo.db.user.find()
         return json_util.dumps(data), 200
     except Exception as e:
-        return jsonify(({ 'msg': f'Error {str(e)}' })), 500
+        return jsonify({ 'msg': f'Error {str(e)}' }), 500
 
 
-def readOneUser(id): #Get user
+def findOneUser(id): #Get user
     try:
         user = mongo.db.user.find_one({'_id': ObjectId(id)})
-        return json_util.dumps(user), 200
+        if user:
+            return json_util.dumps(user), 200
+        else:
+            return jsonify({'msg': 'Resource not found'}), 400
     except Exception as e:
-        return jsonify(({ 'msg': f'Error {str(e)}' })), 500
+        return jsonify({ 'msg': f'Error {str(e)}' }), 500
 
 
-def upgradeUser(id): #Update user
+@jwt_required()
+def editUser(id): #Update user
     try:
-        data = request.get_json()
-        mongo.db.user.update_one({'_id': ObjectId(id)}, {'$set': data})
-        return jsonify({'msg': f'User {id} update successfully'}), 200
+        user = mongo.db.user.find_one({'_id': ObjectId(id)})
+        if user:
+            data = request.get_json()
+            mongo.db.user.update_one({'_id': ObjectId(id)}, {'$set': data})
+            return jsonify({'msg': f'User {id} update successfully'}), 200
+        else:
+            return jsonify({'msg': 'Resource not found'}), 400
     except Exception as e:
-        return jsonify(({ 'msg': f'Error {str(e)}' })), 500
+        return jsonify({ 'msg': f'Error {str(e)}' }), 500
 
 
+@jwt_required()
 def removeUser(id): #Remove user
     try:
-        mongo.db.user.delete_one({'_id': ObjectId(id)})
-        return jsonify({'msg': f'User {id} deleted successfully'}), 200
+        user = mongo.db.user.find_one({'_id': ObjectId(id)})
+        if user:
+            mongo.db.user.delete_one({'_id': ObjectId(id)})
+            return jsonify({'msg': f'User {id} deleted successfully'}), 200
+        else:
+            return jsonify({'msg': 'Resource not found'}), 400
     except Exception as e:
-        return jsonify(({ 'msg': f'Error {str(e)}' })), 500
+        return jsonify({ 'msg': f'Error {str(e)}' }), 500
